Move serial ports to their own utility file
