+++ src/emulator_gpio.c
@@ -30,12 +30,6 @@
 #include "sdr_pin_defines_A0002.h"
 
 /*------------------------------------------------------------------------------
- Globals                                                       
-------------------------------------------------------------------------------*/
-
-extern int serial_port; /* DO NOT MODIFY IN THIS FILE */
-
-/*------------------------------------------------------------------------------
  HAL interfaces                                                       
 ------------------------------------------------------------------------------*/
 
@@ -114,7 +108,7 @@ GPIO_PinState HAL_GPIO_ReadPin(const GPIO_TypeDef *GPIOx, uint16_t GPIO_Pin) {
     if ( ( GPIOx == USB_DETECT_GPIO_PORT )
       && ( GPIO_Pin ==  USB_DETECT_PIN) )
         {
-        return ( serial_port > 0 );
+        return ( emulator_usb_detect( FC_SERIAL_PORT ) );
         }
 
     return GPIO_PIN_RESET;
