+++ src/emulator_ground_station.c
@@ -64,196 +64,9 @@ static uint8_t     buf_ptr = 0;
 
 static pthread_mutex_t state_mutex = PTHREAD_MUTEX_INITIALIZER;
 
-
-/*------------------------------------------------------------------------------
- Static Prototypes                                                     
-------------------------------------------------------------------------------*/
-static USB_STATUS serial_read
-    (
-    void*    rx_data_ptr , /* Buffer to export data to        */
-	size_t   rx_data_size  /* Size of the data to be received */
-    );
-
-static void serial_write
-    (
-    const uint8_t* msg,
-    size_t len
-    );
-
 /*------------------------------------------------------------------------------
  Procedures                                                     
 ------------------------------------------------------------------------------*/
-
-/**
-* @brief Set up the serial connection to SDEC on the GS.                                  
-*/
-bool emulator_prompt_and_open_serial_port_gs
-    (
-    void
-    )
-{
-/* Prompt for serial port */
-char port_buf[12];
-emulator_log("Please enter your serial port in the format /dev/ttyXX or in the format COMX.", EMULATOR_SUBSYSTEM_GS);
-#if defined( _WIN32 ) || defined( __CYGWIN__ )
-#endif
-printf("Input: \n");
-
-if(fgets(port_buf, sizeof(port_buf), stdin) == NULL){
-    emulator_log("Invalid port input.", EMULATOR_SUBSYSTEM_GS);
-    return false;
-}
-
-port_buf[strcspn(port_buf, "\n")] = 0;
-
-char com_buf[4];
-int com_port_num;
-
-strncpy(com_buf, port_buf, 3);
-
-if(strncmp(com_buf, "COM", 3) == 0){
-    sscanf(port_buf+3, "%d", &com_port_num);
-    com_port_num--;
-    uint8_t last_two_digits = com_port_num % 100; /* least significant two digits */
-
-    /* we know this is safe, but we need to ignore the warning */
-    // ETS: THIS IS GROSS. Do not do this.
-    #pragma GCC diagnostic push
-    #pragma GCC diagnostic ignored "-Wformat-truncation"
-    snprintf(port_buf, 12, "/dev/ttyS%u", last_two_digits);
-    #pragma GCC diagnostic pop
-}
-
-serial_port_gs = open(port_buf, O_RDWR | O_NOCTTY | O_NDELAY); // Open the port
-
-if (serial_port_gs < 0) {
-    emulator_log("Error opening serial port.", EMULATOR_SUBSYSTEM_GS);
-    return false;
-}
-
-struct termios tty;
-
-// Read in existing settings, handle errors
-if(tcgetattr(serial_port_gs, &tty) != 0) {
-    emulator_log("tcgetattr failed.", EMULATOR_SUBSYSTEM_GS);
-    return false;
-}
-
-// Configure port settings (baud rate, parity, etc.)
-cfsetospeed(&tty, B921600); // Set output baud rate to 921600
-cfsetispeed(&tty, B921600); // Set input baud rate to 921600
-
-tty.c_cflag &= ~PARENB;        // No parity
-tty.c_cflag &= ~CSTOPB;        // One stop bit
-tty.c_cflag &= ~CSIZE;         // Clear size bits
-tty.c_cflag |= CS8;            // 8 data bits
-tty.c_cflag &= ~CRTSCTS;       // No hardware flow control
-tty.c_cflag |= CREAD | CLOCAL; // Enable reading and ignore modem control lines
-
-tty.c_lflag &= ~ICANON; // Disable canonical mode (line-by-line input)
-tty.c_lflag &= ~ECHO;   // Disable echo
-tty.c_lflag &= ~ECHOE;  // Disable erasure
-tty.c_lflag &= ~ECHONL; // Disable new-line echo
-tty.c_lflag &= ~ISIG;   // Disable interpretation of signal characters
-
-tty.c_iflag &= ~(IXON | IXOFF | IXANY); // Disable software flow control
-tty.c_iflag &= ~(IGNBRK|BRKINT|PARMRK|ISTRIP|INLCR|IGNCR|ICRNL); // Disable special handling of bytes
-
-tty.c_oflag &= ~OPOST; // Prevent special interpretation of output bytes
-tty.c_oflag &= ~ONLCR; // Prevent conversion of newline to carriage return/line feed
-
-tty.c_cc[VTIME] = 1;   // Wait for up to 0.1 seconds (100 ms; 1 decisecond)
-tty.c_cc[VMIN] = 0;    // Minimum number of characters to read
-
-// Save TTY settings, handle errors
-if (tcsetattr(serial_port_gs, TCSANOW, &tty) != 0) {
-    emulator_log("tcsetattr failed.", EMULATOR_SUBSYSTEM_GS);
-    return false;
-}
-
-if (fcntl(serial_port_gs, F_GETFD) == -1) {
-    emulator_log("FD became invalid.", EMULATOR_SUBSYSTEM_GS);
-    return false;
-}
-else {
-    emulator_log("FD remails valid at end of init.", EMULATOR_SUBSYSTEM_GS);
-}
-
-return true;
-
-} /* emulator_prompt_and_open_serial_port_gs */
-
-
-/**
-* @brief Write to the virtual serial port.                                      
-*/
-static void serial_write
-    (
-    const uint8_t* msg,
-    size_t len
-    )
-{
-if ( serial_port_gs < 0 )
-    {
-    return;
-    }
-
-write( serial_port_gs, msg, len );
-
-} /* serial_write */
-
-
-/**
- * @brief Read from the VCP
- * 
- * @param rx_data_ptr The buffer to read to
- * @param rx_data_size The size of the buffer
- *
- * @return The status of the read
- */
-static USB_STATUS serial_read
-    (
-    void*    rx_data_ptr , /* Buffer to export data to        */
-	size_t   rx_data_size  /* Size of the data to be received */
-    )
-{
-
-// Verify fd is still valid before reading
-if (fcntl(serial_port_gs, F_GETFD) == -1) {
-    emulator_log("Read: File descriptor became invalid.", EMULATOR_SUBSYSTEM_GS);
-    return USB_FAIL;
-}
-
-/* Blocking read */
-struct timeval tv;
-tv.tv_sec  = 0;
-tv.tv_usec = 10000; // 10 ms
-
-fd_set rfds;
-FD_ZERO(&rfds);
-FD_SET(serial_port_gs, &rfds);
-
-int ret = select(serial_port_gs + 1, &rfds, NULL, NULL, &tv);
-if (ret < 0) {
-    emulator_log("Select failed.", EMULATOR_SUBSYSTEM_GS);
-    return USB_FAIL;
-} else if (ret == 0) {
-    return USB_TIMEOUT;
-}
-
-memset( rx_data_ptr, 0, rx_data_size );
-int n = read( serial_port_gs, rx_data_ptr, rx_data_size );
-    
-    if (n < 0) {
-        emulator_log("Read failed.", EMULATOR_SUBSYSTEM_GS);
-        return USB_FAIL;
-    } else if (n == 0) {
-        return USB_TIMEOUT;
-    } else {
-        return USB_OK;
-    }
-
-} /* serial_read */
 
 /**
  * @brief Add a new message to the ground station buffer
@@ -310,7 +123,7 @@ while ( 1 )
     {
     memset( rx_buf, 0, 256 );
     memset( tx_buf, 0, 256 );
-    rx_status = serial_read( rx_buf, 1 );
+    rx_status = emulator_serial_read( GS_SERIAL_PORT, rx_buf, 1 );
     
     if( rx_status == USB_FAIL )
         {
@@ -333,7 +146,7 @@ while ( 1 )
             {
             /* Send board identifying code    */
             tx_buf[ 0 ] = GS_HW_CODE;
-            serial_write( tx_buf, 1 );
+            emulator_serial_write( GS_SERIAL_PORT, tx_buf, 1 );
 
             break;
             } /* PING_OP */
@@ -348,7 +161,7 @@ while ( 1 )
             tx_buf[ 1 ] = GS_FW_CODE;
 
             /* Send firmware identifying code */
-            serial_write( tx_buf, 2 );
+            emulator_serial_write( GS_SERIAL_PORT, tx_buf, 2 );
             break;
             } /* CONNECT_OP */
         /*--------------------------------------------------------------
@@ -365,7 +178,7 @@ while ( 1 )
                     {
                     uint8_t len = state_buf[i].len;
                     memcpy( tx_buf, &(state_buf[i].data), len );
-                    serial_write( tx_buf, len );
+                    emulator_serial_write( GS_SERIAL_PORT, tx_buf, len );
                     state_buf[i].timestamp_available = 0xFFFFFFFF;
                     tx_cplt = true;
                     //emulator_debug_logf( "Used an existing message: %d bytes", EMULATOR_SUBSYSTEM_GS, len );
@@ -375,7 +188,7 @@ while ( 1 )
             if( !tx_cplt )
                 {
                 memset( tx_buf, 0, 256 );
-                serial_write( tx_buf, TELEMETRY_MESSAGE_SIZE );
+                emulator_serial_write( GS_SERIAL_PORT, tx_buf, TELEMETRY_MESSAGE_SIZE );
                 //emulator_debug_logf( "Used a blank message: %d bytes", EMULATOR_SUBSYSTEM_GS, TELEMETRY_MESSAGE_SIZE );
                 }
             }
