+++ src/emulator_uart.c
@@ -49,7 +49,6 @@ extern uint8_t            gps_mesg_byte;
 extern uint8_t            rx_buffer[GPSBUFSIZE];
 extern GPS_DATA           gps_data;
 
-int serial_port = -1;
 volatile bool gps_data_it_flag = false;
 
 /* Shared synchronization objects */
@@ -76,17 +75,6 @@ static const char* gps_msgs[] =
 /*------------------------------------------------------------------------------
  Static Prototypes                                                     
 ------------------------------------------------------------------------------*/
-static USB_STATUS serial_read
-    (
-    void*    rx_data_ptr , /* Buffer to export data to        */
-	size_t   rx_data_size  /* Size of the data to be received */
-    );
-
-static void serial_write
-    (
-    const uint8_t* msg,
-    size_t len
-    );
 
 static void gps_read_handler_IT
     (
@@ -100,7 +88,7 @@ static void gps_read_handler_IT
 HAL_StatusTypeDef HAL_UART_Transmit(UART_HandleTypeDef *huart, const uint8_t *pData, uint16_t Size, uint32_t Timeout) {
     if ( huart == &(USB_HUART) )
         {
-        serial_write( pData, (size_t)Size );
+        emulator_serial_write( FC_SERIAL_PORT, pData, (size_t)Size );
         }
     return HAL_OK;
 }
@@ -108,7 +96,7 @@ HAL_StatusTypeDef HAL_UART_Transmit(UART_HandleTypeDef *huart, const uint8_t *pD
 HAL_StatusTypeDef HAL_UART_Transmit_IT(UART_HandleTypeDef *huart, const uint8_t *pData, uint16_t Size) {
     if ( huart == &(USB_HUART) )
         {
-        serial_write( pData, (size_t)Size );
+        emulator_serial_write( FC_SERIAL_PORT, pData, (size_t)Size );
         }
     return HAL_OK;
 }
@@ -116,7 +104,7 @@ HAL_StatusTypeDef HAL_UART_Transmit_IT(UART_HandleTypeDef *huart, const uint8_t 
 HAL_StatusTypeDef HAL_UART_Receive(UART_HandleTypeDef *huart, uint8_t *pData, uint16_t Size, uint32_t Timeout) {
     if ( huart == &(USB_HUART) )
         {
-        serial_read( pData, (size_t)Size );
+        emulator_serial_read( FC_SERIAL_PORT, pData, (size_t)Size );
         }
     return HAL_OK;
 }
@@ -135,173 +123,6 @@ HAL_StatusTypeDef HAL_UART_Receive_IT(UART_HandleTypeDef *huart, uint8_t *pData,
 /*------------------------------------------------------------------------------
  Procedures                                                     
 ------------------------------------------------------------------------------*/
-
-/**
-* Set up the serial connection to SDEC.                                  
-*/
-bool emulator_prompt_and_open_serial_port
-    (
-    void
-    )
-{
-/* Prompt for serial port */
-char port_buf[12];
-emulator_log("Please enter your serial port in the format /dev/ttyXX or in the format COMX.", EMULATOR_SUBSYSTEM_SERIAL);
-#if defined( _WIN32 ) || defined( __CYGWIN__ )
-#endif
-printf("Input: \n");
-
-if(fgets(port_buf, sizeof(port_buf), stdin) == NULL){
-    emulator_log("Invalid port input.", EMULATOR_SUBSYSTEM_SERIAL);
-    return false;
-}
-
-port_buf[strcspn(port_buf, "\n")] = 0;
-
-char com_buf[4];
-int com_port_num;
-
-strncpy(com_buf, port_buf, 3);
-
-if(strncmp(com_buf, "COM", 3) == 0){
-    sscanf(port_buf+3, "%d", &com_port_num);
-    com_port_num--;
-    uint8_t last_two_digits = com_port_num % 100; /* least significant two digits */
-
-    /* we know this is safe, but we need to ignore the warning */
-    // ETS: THIS IS GROSS. Do not do this.
-    #pragma GCC diagnostic push
-    #pragma GCC diagnostic ignored "-Wformat-truncation"
-    snprintf(port_buf, 12, "/dev/ttyS%u", last_two_digits);
-    #pragma GCC diagnostic pop
-}
-
-serial_port = open(port_buf, O_RDWR | O_NOCTTY | O_NDELAY); // Open the port
-
-if (serial_port < 0) {
-    emulator_log("Error opening serial port.", EMULATOR_SUBSYSTEM_SERIAL);
-    return false;
-}
-
-struct termios tty;
-
-// Read in existing settings, handle errors
-if(tcgetattr(serial_port, &tty) != 0) {
-    emulator_log("tcgetattr failed.", EMULATOR_SUBSYSTEM_SERIAL);
-    return false;
-}
-
-// Configure port settings (baud rate, parity, etc.)
-cfsetospeed(&tty, B921600); // Set output baud rate to 921600
-cfsetispeed(&tty, B921600); // Set input baud rate to 921600
-
-tty.c_cflag &= ~PARENB;        // No parity
-tty.c_cflag &= ~CSTOPB;        // One stop bit
-tty.c_cflag &= ~CSIZE;         // Clear size bits
-tty.c_cflag |= CS8;            // 8 data bits
-tty.c_cflag &= ~CRTSCTS;       // No hardware flow control
-tty.c_cflag |= CREAD | CLOCAL; // Enable reading and ignore modem control lines
-
-tty.c_lflag &= ~ICANON; // Disable canonical mode (line-by-line input)
-tty.c_lflag &= ~ECHO;   // Disable echo
-tty.c_lflag &= ~ECHOE;  // Disable erasure
-tty.c_lflag &= ~ECHONL; // Disable new-line echo
-tty.c_lflag &= ~ISIG;   // Disable interpretation of signal characters
-
-tty.c_iflag &= ~(IXON | IXOFF | IXANY); // Disable software flow control
-tty.c_iflag &= ~(IGNBRK|BRKINT|PARMRK|ISTRIP|INLCR|IGNCR|ICRNL); // Disable special handling of bytes
-
-tty.c_oflag &= ~OPOST; // Prevent special interpretation of output bytes
-tty.c_oflag &= ~ONLCR; // Prevent conversion of newline to carriage return/line feed
-
-tty.c_cc[VTIME] = 1;   // Wait for up to 0.1 seconds (100 ms; 1 decisecond)
-tty.c_cc[VMIN] = 0;    // Minimum number of characters to read
-
-// Save TTY settings, handle errors
-if (tcsetattr(serial_port, TCSANOW, &tty) != 0) {
-    emulator_log("tcsetattr failed.", EMULATOR_SUBSYSTEM_SERIAL);
-    return false;
-}
-
-if (fcntl(serial_port, F_GETFD) == -1) {
-    emulator_log("FD became invalid.", EMULATOR_SUBSYSTEM_SERIAL);
-    return false;
-}
-else {
-    emulator_log("FD remails valid at end of init.", EMULATOR_SUBSYSTEM_SERIAL);
-}
-
-return true;
-
-} /* emulator_prompt_and_open_serial_port */
-
-
-/**
-* Write to the virtual serial port.                                      
-*/
-static void serial_write
-    (
-    const uint8_t* msg,
-    size_t len
-    )
-{
-if ( serial_port < 0 )
-    {
-    return;
-    }
-
-write( serial_port, msg, len );
-
-} /* serial_write */
-
-
-/**
-* Read from the virtual serial port.                                     
-*/
-static USB_STATUS serial_read
-    (
-    void*    rx_data_ptr , /* Buffer to export data to        */
-	size_t   rx_data_size  /* Size of the data to be received */
-    )
-{
-
-// Verify fd is still valid before reading
-if (fcntl(serial_port, F_GETFD) == -1) {
-    emulator_log("Read: File descriptor became invalid.", EMULATOR_SUBSYSTEM_SERIAL);
-    return USB_FAIL;
-}
-
-/* Blocking read */
-struct timeval tv;
-tv.tv_sec  = 0;
-tv.tv_usec = 10000; // 10 ms
-
-fd_set rfds;
-FD_ZERO(&rfds);
-FD_SET(serial_port, &rfds);
-
-int ret = select(serial_port + 1, &rfds, NULL, NULL, &tv);
-if (ret < 0) {
-    emulator_log("Select failed.", EMULATOR_SUBSYSTEM_SERIAL);
-    return USB_FAIL;
-} else if (ret == 0) {
-    return USB_TIMEOUT;
-}
-
-memset( rx_data_ptr, 0, rx_data_size );
-int n = read( serial_port, rx_data_ptr, rx_data_size );
-    
-    if (n < 0) {
-        emulator_log("Read failed.", EMULATOR_SUBSYSTEM_SERIAL);
-        return USB_FAIL;
-    } else if (n == 0) {
-        return USB_TIMEOUT;
-    } else {
-        return USB_OK;
-    }
-
-} /* serial_read */
-
 
 /**
 * Listen for and fulfill UART GPS IT I/O.
